% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runes.R
\name{runes}
\alias{runes}
\title{Elder Futhark Runes in \code{R}}
\usage{
runes(x)
}
\arguments{
\item{x}{a character string}
}
\value{
If input is a single character string then output will be a character
string with unicode Elder Futhark runes replacing the 'A-Z' characters except
'q' and 'x'
}
\description{
\code{runes()} returns the input string argument with all 'A-Z' characters
replaced by Elder Futhark Runes. You can use this in console output, scripts,
RMarkdown documents, anything with Unicode/UTF-8 support
}
\examples{
cat(runes("Your Results Are: "), 7 * 49, "\n")

runes("hello world")
}
