% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runes_table.R
\name{runes_table}
\alias{runes_table}
\title{runes_table()}
\usage{
runes_table(futhark = "elder", branch = NA)
}
\arguments{
\item{futhark}{which futhark set you'd like, defualts to "elder" and the
argument must be a string}

\item{branch}{if futhark is "younger" then you must also choose wether you
would like short or long branch younger futhark}
}
\description{
\code{runes_table()} returns a data.frame with 3 variable text columns.AX
}
\details{
The data.frame returned by this function returns 3 columns
\enumerate{
\item The unicode sequence that represents the rune character.
\item The English a-z characters of the alphabet
\item The rune itself showing how it would also be displayed.
}
}
\examples{
runes_table()

}
