% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runes_table.R
\name{runes_table}
\alias{runes_table}
\title{runes_table()}
\usage{
runes_table()
}
\description{
\code{runes_table()} returns a data.frame with 3 variable text columns.AX
}
\details{
The data.frame returned by this funtion returns 3 columns
\enumerate{
\item The unicode characer that represents the rune character.
\item The English a-z characters of the alhpabet
\item The rune itself showing how it would also be displayed.
}
}
\examples{
runes_table()

}
